Factor out event scraping
